Lower compile-times of compiling `objc2` crates
By disabling default features, and only selectively enabling what we
need. This was done previously, but was removed in fae2979, probably to
make iteration easier. But since native-dialog is a library, we should
probably cater towards users in this regard. For example, when building
Veloren, building objc2-app-kit with all features takes up ~16% of the
entire build.