win: Include `hresult` when printing "`Other`" error
The `error_method` field, as its name "somewhat" describes, is the
**name** of the COM method that failed with a particular error code.
The actual error code is in the `hresult` field.

To make this error more easily readable to users, wrap it in
`std::io::Error` using `from_raw_os_error()` which takes `HRESULT`
integers.

Note that like the other `#[from]` (which implies `#[source]`) errors
above in this crate these nested errors are *not* included in the
`Display` formatter generated by `thiserror`; callers are expected to
`Debug`-print the returned `Error` or wrap it in a more useful type like
`anyhow::Error` which nicely `Debug`-prints all `.source()` errors as a
stack (see https://github.com/dtolnay/thiserror/issues/98 to hopefully
have the same behaviour in `thiserror` some day).